remove logic to decidim-core
