+++ decidim-core/app/packs/entrypoints/decidim_geocoding.js
@@ -1,0 +1,2 @@
+import "src/decidim/geocoding/reverse_geocoding.js"
+import "stylesheets/decidim/geocoding_addons.scss";
