+++ decidim-core/app/packs/stylesheets/decidim/geocoding_addons.scss
