+++ decidim-core/lib/decidim/core.rb
@@ -845,6 +845,12 @@ module Decidim
     0.seconds
   end
 
+  # Public Setting that allows to configure which component will have "Use my current location" button
+  # in a geocoded address field. Accepts an array of component manifest names
+  config_accessor :show_my_location_button do
+    [:proposals]
+  end
+
   def self.machine_translation_service_klass
     return unless Decidim.enable_machine_translations
