+++ decidim-core/lib/decidim/map/autocomplete.rb
@@ -11,23 +11,31 @@ module Decidim
       class Builder < Decidim::Map::Frontend::Builder
         delegate :current_component, :content_tag, :asset_pack_path, to: :template
 
-        # Displays the geocoding field element's markup for the view.
+        # Renders the geocoding field element for the view. If the component supports
+        # geolocation (as determined by `show_my_location_button?`), the field will include
+        # a button allowing the user to use their current location to autofill the address.
+        # Otherwise, a standard geocoding input field is rendered without the location button.
         #
-        # @param object_name [String, Symbol] The name for the object for which
-        #   the field is generated for.
-        # @param method [String, Symbol] The method/property in the object that
-        #   the field is for.
-        # @param options [Hash] Extra options for the field.
-        # @return [String] The field element's markup.
+        # @param object_name [String, Symbol] The name of the object the field is being generated for.
+        # @param method [String, Symbol] The specific attribute or property of the object.
+        # @param options [Hash] Additional options for customizing the field's behavior and appearance.
+        # @return [String] The HTML markup for the geocoding field.
         def geocoding_field(object_name, method, options = {})
-          return original_geocoding_field(object_name, method, options) unless show_my_location_button?
-
-          append_assets
-          template.snippets.add(:decidim_proposals_geocoding_scripts, template.append_javascript_pack_tag("decidim_proposals_geocoding"))
-          template.snippets.add(:decidim_proposals_geocoding_styles, template.append_stylesheet_pack_tag("decidim_proposals_geocoding"))
-
           options[:autocomplete] ||= "off"
-          options[:class] ||= "input-group-field"
+          append_assets
+
+          if show_my_location_button?
+            geocoding_field_with_location_button(object_name, method, options)
+          else
+            geocoding_field_without_location_button(object_name, method, options)
+          end
+        end
+
+        private
+
+        def geocoding_field_with_location_button(object_name, method, options)
+          template.snippets.add(:decidim_geocoding_scripts, template.append_javascript_pack_tag("decidim_geocoding"))
+          template.snippets.add(:decidim_geocoding_styles, template.append_stylesheet_pack_tag("decidim_geocoding"))
 
           template.content_tag(:div, class: "geocoding-container") do
             template.text_field(
@@ -42,7 +50,7 @@ module Decidim
                                        longitude: "#{object_name}_longitude",
                                        error_no_location: I18n.t("errors.no_device_location", scope: "decidim.proposals.forms"),
                                        error_unsupported: I18n.t("errors.device_not_supported", scope: "decidim.proposals.forms"),
-                                       url: template.url_for(controller: "geolocation", action: "locate")
+                                       url: Decidim::Core::Engine.routes.url_helpers.locate_path
                                      }) do
                   template.icon("map-pin-line", role: "img", "aria-hidden": true) + " #{I18n.t("use_my_location", scope: "decidim.proposals.forms")}"
                 end
@@ -50,14 +58,18 @@ module Decidim
           end
         end
 
-        alias original_geocoding_field geocoding_field
-
-        private
+        def geocoding_field_without_location_button(object_name, method, options)
+          template.text_field(
+            object_name,
+            method,
+            options.merge("data-decidim-geocoding" => view_options.to_json)
+          )
+        end
 
         def show_my_location_button?
           return unless template.respond_to?(:current_component)
 
-          Decidim::Proposals.show_my_location_button.include?(template.current_component.manifest_name.to_sym)
+          Decidim.show_my_location_button.include?(template.current_component.manifest_name.to_sym)
         end
       end
