--- decidim-proposals/app/controllers/decidim/proposals/admin/geolocation_controller.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module Proposals
-    module Admin
-      class GeolocationController < Decidim::Proposals::GeolocationController
-      end
-    end
-  end
-end
