+++ decidim-proposals/lib/decidim/proposals.rb
@@ -34,12 +34,6 @@ module Decidim
       3
     end
 
-    # Public Setting that allows to configure which component will have "Use my current location" button
-    # in a geocoded address field. Accepts an array of component manifest names
-    config_accessor :show_my_location_button do
-      [:proposals]
-    end
-
     def self.proposal_states_colors
       {
         gray: {
