+++ decidim-proposals/lib/decidim/proposals/engine.rb
@@ -28,9 +28,6 @@ module Decidim
             end
           end
         end
-
-        post :locate, to: "geolocation#locate"
-
         resources :collaborative_drafts, except: [:destroy] do
           member do
             post :request_access, controller: "collaborative_draft_collaborator_requests"
